Edit Selesai

+++ database/migrations/2023_02_21_144258_create_stoks_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
             $table->integer('stokAwal')->nullable();
             $table->integer('stokAkhir')->nullable();
             $table->integer('terjual')->nullable();
-            $table->float('total')->nullable();
+            $table->integer('total')->nullable();
